Rotate the new isup tables as well
Similar to sip_capture_call rotate (and partition) the isup_capture_all
tables. Add a second template for the ISUP table to the script. It is
not very object orientated but the smallest change I can think of. Given
the expected number of rows it makes sense to treat it like the
sip_capture_call tables.