isup: Take src-ip/dest-ip into account when creating unique messages
If we capture on STP A and STP B and a message is being forwarded from
A to B and then B to C we want to see the message twice. The current
unique feature would limit it to only A to B. Put the source and the
destination ip into the unique_key.